backend/requirements.txt
django
djangorestframework
gunicorn
psycopg2-binary
djangorestframework-simplejwt
django-ratelimit
prometheus_client
transformers
torch
sentence-transformers
faiss-cpu
accelerate>=0.26.0
black
pre-commit
drf-spectacular
drf-spectacular==0.28.0
openai
dotenv
openai>=1.0.0
black
flake8
sentence-transformers
faiss-cpu
numpy
PyMuPDF



django-cors-headers==4.3.1
django-filter==23.5
